\name{cas2015}
\alias{cas2015}
\docType{data}
\title{
CAS2015 Dataset
}
\description{This dataset was created by Braga, A. C. and allows the comparison of two independent samples.}

\usage{data(cas2015)}
\format{
  A data frame with a total of 800 observations on the following 2 variables and respectives status.
  \describe{
    \item{\code{mod1}}{CRIBM}
    \item{\code{status1}}{Result1}
    \item{\code{mod2}}{CRIBF}
    \item{\code{status2}}{Result2}
  }
}
\details{
The dataset contains the values of the indicator (CRIB) for 2 different groups (sex: M/F) and respective results, from 0 (alive) to 1 (deceased). These samples are unpaired, therefore presenting different statuses for each one.
}
\source{
COELHO, S. AND BRAGA, A. C.: Performance Evaluation of Two Software for Analysis Through ROC Curves: Comp2ROC vs SPSS. Computational Science and Its Applications -- ICCSA 2015; p. 144-156; Springer International Publishing., ISBN: 978-3-319-21406-1.
}

\keyword{datasets}
