\name{curvesegslope}
\alias{curvesegslope}
\title{Segment Slopes}

\description{
This function allows to calculate the ROC curve segments slope through the points that are given by parameter.
}
\usage{
curvesegslope(curve.fpr, curve.tpr)
}

\arguments{
  \item{curve.fpr}{
False positive rate vector with all points of the given Curve
}
  \item{curve.tpr}{
True positive rate vector with all points of the given Curve
}
}

\value{
This function returns a vector with all segments slopes 

}

\keyword{ Curve }
\keyword{ Slope }
\keyword{ ROC }
\keyword{ Segment }
