\name{lineslope}
\alias{lineslope}
\title{Sampling Lines Slope}

\description{
This function allows to calculate the sample lines slope that were drawn beginning at the reference point.
}

\usage{
lineslope(K)
}

\arguments{
\item{K}{
Number of sampling lines that we want to create
}
}

\value{
This function returns a vector with all slopes of the sampling lines that we create
}

\examples{
K=100
lineslope(K)
}

\keyword{ Sampling }
\keyword{ Line }

