\name{rocsampling.summary}
\alias{rocsampling.summary}
\title{Summary of ROC Sampling}

\description{
This function allows to see with a simple interface the results obtained in \code{rocsampling}.
}
\usage{
rocsampling.summary(result, mod1, mod2)
}

\arguments{
  \item{result}{
List with results obtained throught the use of  \code{rocsampling}
}
  \item{mod1}{
Name of the column of dataframe that represents the first modality
}
  \item{mod2}{
Name of the column of dataframe that represents the second modality
}
}

\seealso{
\code{\link{rocsampling}}
}

\keyword{ Summary }
\keyword{ ROC }
\keyword{ Sampling }
