\name{zhang}
\alias{zhang}
\docType{data}
\title{
Zhang Dataset
}
\description{This dataset was created by Zhang and we use it as example on our package}

\usage{data(zhang)}
\format{
  A data frame with 2410 observations on the following 3 variables.
  \describe{
    \item{\code{mod1}}{modality 1}
    \item{\code{status}}{status}
    \item{\code{mod2}}{modality 2}
  }
}
\details{
This modalities are related to each other, so they have the same status}
\source{
ZHANG, D. AND ZHOU, X.AND FREEMAN, D. AND FREEMAN, J. 2002. A nonparametric method for the comparison of partial areas under ROC curves and its application to large health care data sets In Stat. Med., Vol. 21 N. 5 701-715.
}

\keyword{datasets}
